Moved validator-related tests to ConfigOptionsValidatorsTraitTest
