+++ test/Config/Plugin/UrlShortenerConfigCustomizerTest.php
@@ -8,7 +8,6 @@ use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
 use Shlinkio\Shlink\Installer\Config\Plugin\UrlShortenerConfigCustomizer;
-use Shlinkio\Shlink\Installer\Exception\InvalidConfigOptionException;
 use Shlinkio\Shlink\Installer\Model\CustomizableAppConfig;
 use ShlinkioTest\Shlink\Installer\Util\TestUtilsTrait;
 use Symfony\Component\Console\Style\SymfonyStyle;
@@ -114,43 +113,5 @@ class UrlShortenerConfigCustomizerTest extends TestCase
             UrlShortenerConfigCustomizer::NOTIFY_VISITS_WEBHOOKS => true,
             UrlShortenerConfigCustomizer::VISITS_WEBHOOKS => 'webhook',
         ]];
-    }
-
-    /**
-     * @test
-     * @dataProvider provideWebhooks
-     */
-    public function webhooksAreProperlySplitAndValidated(?string $webhooks, array $expectedResult): void
-    {
-        $result = $this->plugin->splitAndValidateMultipleUrls($webhooks);
-        $this->assertEquals($expectedResult, $result);
-    }
-
-    public function provideWebhooks(): iterable
-    {
-        yield 'no webhooks' => [null, []];
-        yield 'single webhook' => ['https://foo.com/bar', ['https://foo.com/bar']];
-        yield 'multiple webhook' => ['https://foo.com/bar,http://bar.io/foo/bar', [
-            'https://foo.com/bar',
-            'http://bar.io/foo/bar',
-        ]];
-    }
-
-    /**
-     * @test
-     * @dataProvider provideInvalidWebhooks
-     */
-    public function webhooksFailWhenProvidedValueIsNotValidUrl(string $webhooks): void
-    {
-        $this->expectException(InvalidConfigOptionException::class);
-        $this->plugin->splitAndValidateMultipleUrls($webhooks);
-    }
-
-    public function provideInvalidWebhooks(): iterable
-    {
-        yield 'single invalid webhook' => ['invalid'];
-        yield 'first invalid webhook' => ['invalid,http://bar.io/foo/bar'];
-        yield 'last invalid webhook' => ['http://bar.io/foo/bar,invalid'];
-        yield 'middle invalid webhook' => ['http://bar.io/foo/bar,invalid,https://foo.com/bar'];
     }
 }
